Redirect Derek file to openfaas/faas file
This change redirects the derek file to the main openfaas/faas
repo to allow us to maintain a single source of authorized derek
users at the main repository.

The OpenFaas organization uses several repositories for the code
base that are all related. Consolidating the derek file will make
it easier to maintain when adding or removing users

Signed-off-by: Burton Rheutan <rheutan7@gmail.com>