Fix java8 template to support resources
With this change all resources within function's
`src/main/resources` will be added to function.jar. Before that
they used to be omited. The distribution, created after `gradle
build` was modifying the jars, excluding all resources. Fixed
that by deleting this distribution and creating a controlled one
with `gradle distZip`

Signed-off-by: Ivana Yovcheva (VMware) <iyovcheva@vmware.com>