Update the java8 template to honor cgroups memory limits
This commit upgrades the openjdk version to the earliest that supports
cgroup memory limits. It also applies the JVM arguments required to
enable the support.

This change is needed for the JVM to recognize the actual memory
made available to it via the container instead of believing it has
access to the amount of memory possessed by the node running the
container.

Signed-off-by: Chad Taylor <taylor.thomas.c@gmail.com>