Add source compatibility flags
Pin the source compatibility and target compatibility to java 1.8
because that is the intended runtime. This work is part of #137.

I added the flags to the subprojects block, and the function
build.gradle explicitly. The reason for the extra block in the function
is because developers of java functions mainly only be working with this
build.gradle, and the rest of the gradle is under the covers during the
function build process that the faas-cli provides.

Signed-off-by: devongleeson <devon.gleeson@gmail.com>