Add .npmrc files to node templates
This change adds an empty .npmrc file to the node templates
as well as adding the NPM_TOKEN build arg to allow users
to specify a private npm registry, or pass credentials to the
templates during build process.

Use of the NPM_TOKEN build arg means that the credentials are
not stored with the images and remain secure once the function
is built and deployed.

Signed-off-by: Burton Rheutan <rheutan7@gmail.com>