Remove Java8 and arm-based templates
Users are still using the java8 template despite java8 being
deprecated. This should prevent that going forward, the java11
template is compatible.

Users can use an older SHA / tag of the repo if they are still
migrating off the deprecated java8 template:

https://github.com/openfaas/templates/releases/tag/1.9.0

The armhf and arm64 templates are no longer required, since
they are duplicates of the normal templates. What matters more
for arm is using a multi-arch template, or building on the
correct device. There's no explicit need to use the armhf
template as it adds nothing.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>