Fix body parsing for x-www-form-urlencoded and empty
Form data posted with application/x-www-form-urlencoded
was ignored. This has been addressed and tested, and the body
is now populated.

The default handler now shows the content type received so
that users can understand what they are sending in via curl
or other mechanisms.

When no content-type was given, the template did not attempt to
do any form of parsing. This has been updated, so set a plain
text body type when the sender doesn't provide one.

Tested with node and express.js with various content-types.

A tech-debt item is that the body parser is now considered
deprecated, and a new parser needs to be moved to.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>