Update go version to 1.16 for alpine image
We need to bump up go version by 1 for alpine image
for compatibility and use alpine 3.13.
At the time of this writing,
latest go version is 1.17 and latest alpine is 3.14.

Signed-off-by: F. Talha Altinel <talhaaltinel@hotmail.com>

Update go version to 1.16 for alpine image

We need to bump up go version by 1 for alpine image
for compatibility and use alpine 3.13.
At the time of this writing,
latest go version is 1.17 and latest alpine is 3.14.

Signed-off-by: F. Talha Altinel <talhaaltinel@hotmail.com>